flask/requirements.txt
apscheduler>=3.10.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
flask_cors==4.0.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
pygeohash==1.2.0
python-dotenv>=1.0.0
Zenora>=0.0.3.post1
gunicorn
gevent
gevent-websocket